+++ java/voronoi/Point.java
@@ -5,20 +5,20 @@ package voronoi;
  */
 public class Point implements Comparable<Point>
 {
-	private int x, y;
+	private double x, y;
 
-	public Point(int x, int y)
+	public Point(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX()
+	public double getX()
 	{
 		return x;
 	}
 
-	public int getY()
+	public double getY()
 	{
 		return y;
 	}
@@ -26,8 +26,8 @@ public class Point implements Comparable<Point>
 	@Override
 	public int compareTo(Point point)
 	{
-		int compareY = Integer.compare(this.y, point.y);
-		return compareY == 0 ? Integer.compare(this.x, point.x) : compareY;
+		int compareY = Double.compare(this.y, point.y);
+		return compareY == 0 ? Double.compare(this.x, point.x) : compareY;
 	}
 
 	@Override
